requirements.txt
psutil~=5.7.0
Flask~=1.1.2
discord.py
colorama~=0.4.3
pymongo>=3.10.1